requirements.txt
gunicorn==20.1.0
dj-database-url==0.5.0
Django==4.0
django-countries==7.2.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.0
Markdown==3.3.6
psycopg2==2.9.2
requests==2.22.0
requests-unixsocket==0.2.0
simplejson==3.16.0
sqlparse==0.4.2
urllib3==1.25.8
wadllib==1.3.3
whitenoise==5.3.0
